[AOS] 바텀 시트 사용자 선택 데이터 동기화 (#342)
* [AOS] fix: 리뷰 내용 반영 #311

* [AOS] fix: 색상 선택 동적 마진 값 #311

* [AOS] feat: 바텀 시트 데이터 연동 #311

* [AOS] chroe: 리소스 파일 추가 및 불 필요한 파일 삭제 #311

* [AOS] chroe: 불 필요한 파일 정리 및 변수 타입 변경#311

* [AOS] feat: 사용자 선택 데이터, 바텀 시트 바인딩 #341

+++ Android/app/src/main/java/org/softeer_2nd/caArt/model/data/Engine.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class Engine(
     @SerializedName("engineName") override val itemName: String,
     @SerializedName("engineImage") override val imageUrl: String,
-    @SerializedName("enginePrice") val enginePrice: Int,
+    @SerializedName("enginePrice") val enginePrice: Long,
     override val description: String,
     override val summary: String,
     override val maxPower: String,
